put server side trpc code in a single file
